Create SQLCloseable interface
The main reason of this is because I use transactions, in fact, I declare my connection and prepared statement into a try-with-resources statement, but I dont have access to that connection outside the try, so when a problem occurs, I can't roll back the connection into the catch.

I can roll back into the try-with-resources statement with AutoCloseable, but the problem is that AutoCloseable throws an Exception when I want an SQLException.

Finally, I have to create an interface who extends AutoCloseable for overriding the close method with a throws SQLException.